+++ tests/test_beta.py
@@ -6,7 +6,6 @@ from typing import Dict, Optional
 
 sys.path.insert(0, os.path.abspath(os.path.join(os.path.dirname(__file__), '..')))
 
-
 from src.core import (
     CachedClientSession,
     CachedCurlCffiSession,
@@ -55,6 +54,7 @@ class Bot:
             self.load_scanlators(scanlaors)
 
     async def close(self):
+        await self.db.conn.close()
         await self.curl_session.close()
         await self.session.close()
 
@@ -122,24 +122,33 @@ async def main():
 
     async with Bot(proxy_url=proxy_url) as bot:
         init_scanlators(bot, scanlators)
-        key = "asura"
-        url = "https://asuracomic.net/series/return-of-the-sss-class-ranker-e03dfc4f"
+        key = "comick"
+        url = "https://api.comick.fun/comic/05-one-punch-man"
         query = "he"
         scanlator = scanlators[key]
-        title = await scanlator.get_title(url)
+        title = await scanlator.get_title(url);
+        # print("Title:", title)
+        _id = await scanlator.get_id(url);
+        # print("ID:", _id)
+        all_chapters = await scanlator.get_all_chapters(url);
+        # print("All chapters:", all_chapters)
+        status = await scanlator.get_status(url);
+        # print("Status:", status)
+        synopsis = await scanlator.get_synopsis(url);
+        # print("Synopsis:", synopsis)
+        cover = await scanlator.get_cover(url);
+        # print("Cover:", cover)
+        fp_manga = await scanlator.get_fp_partial_manga();
+        # print("FP Manhwa:", fp_manga)
+        search_result = await scanlator.search(query);
+        # print("Search:", search_result)
+        manga = await scanlator.make_manga_object(url, load_from_db=False)
+        # print("Manga obj:", manga)
+        updates_result = await scanlator.check_updates(manga);
+        # print("Updates result:", updates_result)
 
-        _id = await scanlator.get_id(url)
-        all_chapters = await scanlator.get_all_chapters(url)
-        status = await scanlator.get_status(url)
-        synopsis = await scanlator.get_synopsis(url)
-        cover = await scanlator.get_cover(url)
-        fp_manga = await scanlator.get_fp_partial_manga()
-        search_result = await scanlator.search(query)
-        manga = await scanlator.make_manga_object(url)
-        updates_result = await scanlator.check_updates(manga)
-        #
         results = (
-            f"Title: {title}", f"ID: {_id}", f"All Chapters: {all_chapters}", f"Status: {status}",
+            f"Title: {title}", f"ID: {_id}", f"URL: {manga.url}", f"All Chapters: {all_chapters}", f"Status: {status}",
             f"Synopsis: {synopsis}", f"Cover: {cover}", f"FP Manhwa: {fp_manga}", f"Search: {search_result}",
             f"Manga obj: {manga}", f"Updates result: {updates_result}"
         )
@@ -170,4 +179,9 @@ if __name__ == "__main__":
             "filelock"
         ]
     )
-    asyncio.run(main())
+    try:
+        loop = asyncio.new_event_loop()
+        asyncio.set_event_loop(loop)
+        loop.run_until_complete(main())
+    finally:
+        loop.close()